Add Storage class for more OOP
